Added API Endpoint: Fetching a single event based on its UUID
Every single event inserted in the database will have a server-generated UUID. This endpoint allows the client to access information about a specific event based on this UUID. For now, the information sent back from this endpoint is identical to the information sent back by the global event fetching endpoint. In the future, however, this new endpoint will provide more detailed information than the global one.